+++ include/structs.h
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/21 08:37:38 by jkauker           #+#    #+#             */
-/*   Updated: 2024/06/29 19:37:53 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/29 20:02:27 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -157,6 +157,8 @@ typedef struct s_player
 {
 	t_transform	transform;
 	t_transform	spawn_transform;
+
+	int			elemental_type;
 
 	int			health;